Small Compile issue

+++ sw/airborne/modules/ATMOS/multiGain.h
@@ -21,7 +21,7 @@ extern void SetGainSetA(void);
 extern void SetGainSetB(void);
 extern void SetGainSetC(void);
 
-extern void SetGainPercentage(uint8_t percent);
+extern void SetGainPercentage(uint32_t percent);
 
 extern void multiGain_SetGainSetHandler(int8_t v);
 extern void transveh_multigain_setGainSet(struct Int32AttitudeGains * gainSet);
